Refresh template with `bindings/c/iree/compiler/` code. (#1)
The new `hello_compiler/hello_compiler.c` file is a minimal example of
using the compiler C API and loader. The README has instructions for
using either a source build of the compiler or a binary release from a
pip package.

I think it would be useful to add subdirectories for each compiler
extension mechanism (plugins, ` --iree-hal-preprocess-executables-with`,
etc.). For now, this tries to follow the structure of
https://github.com/iree-org/iree-template-runtime-cmake.